+++ packages/common/src/utils/NetworkUtil.ts
@@ -1,0 +1,7 @@
+import type { CaipNetworkId } from '@web3modal/core'
+
+export const NetworkUtil = {
+  caipNetworkIdToNumber(caipnetworkId?: CaipNetworkId) {
+    return caipnetworkId ? Number(caipnetworkId.split(':')[1]) : undefined
+  }
+}