+++ packages/scaffold/src/partials/w3m-account-wallet-features-widget/styles.ts
@@ -11,6 +11,10 @@ export default css`
   }
 
   wui-profile-button {
+    margin-top: calc(-1 * var(--wui-spacing-2l));
+  }
+
+  wui-promo + wui-profile-button {
     margin-top: var(--wui-spacing-2l);
   }