+++ packages/core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @web3modal/core
 
+## 4.1.3-5f2ae345.1
+
+### Patch Changes
+
+- canary: test imports
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.3-5f2ae345.1
+  - @web3modal/wallet@4.1.3-5f2ae345.1
+
+## 4.1.3-8e039e.0
+
+### Patch Changes
+
+- feat: update optional dependencies
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.3-8e039e.0
+  - @web3modal/wallet@4.1.3-8e039e.0
+
+## 4.1.2
+
+### Patch Changes
+
+- 4.1.2 release
+
+- Updated dependencies []:
+  - @web3modal/common@4.1.2 - @web3modal/wallet@4.1.2
+
 ## 4.2.0-4b5257b4.1
 
 ### Minor Changes